pyproject.toml
[project]
name = "try-agent-llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.9.6",
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "llama-index-core>=0.14.15",
    "llama-index-embeddings-google-genai>=0.3.2",
    "llama-index-llms-google-genai>=0.8.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
